ui/bug: add the case when there is no PV
Refs: #1673
