+++ ui/src/containers/NodeVolumes.js
@@ -13,6 +13,7 @@ import {
   sortCapacity,
   useRefreshEffect,
 } from '../services/utils';
+import { isVolumeDeletable } from '../services/NodeVolumesUtils';
 import {
   refreshVolumesAction,
   stopRefreshVolumesAction,
@@ -27,7 +28,6 @@ import {
   STATUS_FAILED,
   STATUS_AVAILABLE,
   STATUS_BOUND,
-  STATUS_RELEASED,
 } from '../constants';
 
 const VolumeTable = styled.div`
@@ -119,48 +119,6 @@ const NodeVolumes = props => {
     setSortDirection(sortDirection);
   };
 
-  const isVolumeDeletable = rowData => {
-    const volumeStatus = rowData.status;
-    const volumeName = rowData.name;
-
-    switch (volumeStatus) {
-      case STATUS_UNKNOWN:
-      case STATUS_PENDING:
-      case STATUS_TERMINATING:
-        return false;
-      case STATUS_FAILED:
-      case STATUS_AVAILABLE:
-        if (persistentVolumes?.length === 0) {
-          return true;
-        } else {
-          const persistentVolume = persistentVolumes.find(
-            pv => pv?.metadata?.name === volumeName,
-          );
-          const persistentVolumeStatus = persistentVolume?.status?.phase;
-
-          switch (persistentVolumeStatus) {
-            case STATUS_FAILED:
-            case STATUS_AVAILABLE:
-            case STATUS_RELEASED:
-              return true;
-            case STATUS_PENDING:
-            case STATUS_BOUND:
-              return false;
-            default:
-              console.error(
-                `Unexpected state for PersistentVolume ${volumeName}:${persistentVolumeStatus}`,
-              );
-              return false;
-          }
-        }
-      default:
-        console.error(
-          `Unexpected state for Volume ${volumeName}:${volumeStatus}`,
-        );
-        return false;
-    }
-  };
-
   const columns = [
     {
       label: intl.messages.name,
@@ -204,7 +162,7 @@ const NodeVolumes = props => {
       disableSort: true,
       width: 80,
       renderer: (data, rowData) => {
-        const isEnableClick = isVolumeDeletable(rowData);
+        const isEnableClick = isVolumeDeletable(rowData, persistentVolumes);
         const isTriggerTooltip = () => {
           if (isEnableClick) {
             return '';
